+++ src/main/kotlin/cloud/glitchdev/rfu/config/categories/GeneralFishing.kt
@@ -68,6 +68,11 @@ object GeneralFishing : Category("General Fishing") {
         condition = { bossHealthBars }
     }
 
+    var boostPollingRate by boolean(true) {
+        name = Literal("Boost Polling Rate")
+        description = Literal("Makes detections more frequent when the health bar is active (Probably wont but may cause lag, hence the option)")
+    }
+
     val HEALTH_BAR_REGEX
         get() = healthBarMobs.joinToString("|").toExactRegex()